requirements.txt
matplotlib
numpy 
transformers
pillow
qwen-vl-utils[decord]==0.0.8
gradio
sentence-transformers
TTS